approach2_exp/requirements.txt
absl-py==1.4.0
accelerate==0.18.0
adatest==0.3.5
aiohttp==3.8.4
aiohttp-security==0.4.0
aiohttp-session==2.12.0
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.1
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bitsandbytes==0.38.1
blinker==1.6.2
Brlapi==0.7.0
bs4==0.0.1
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
chromadb==0.4.15
clarifai==9.7.6
clarifai-grpc==9.7.4
cleo==2.0.1
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.3
cohere==4.21
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
command-not-found==0.3
contourpy==1.0.7
crashtest==0.4.1
cryptography==40.0.2
cupshelpers==1.0
cycler==0.11.0
dataclasses-json==0.5.14
datasets==2.12.0
dbus-python==1.2.16
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.14
diffusers==0.24.0
dill==0.3.6
diskcache==5.6.1
distlib==0.3.7
distro==1.8.0
distro-info==0.23+ubuntu1.1
Django==4.2.5
django-cors-headers==4.3.1
dnspython==2.4.2
docker-pycreds==0.4.0
docopt==0.6.2
duplicity==0.8.12.0
effdet==0.4.1
einops==0.7.0
emoji==2.8.0
entrypoints==0.3
et-xmlfile==1.1.0
evaluate==0.4.0
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.104.1
fastavro==1.8.2
fasteners==0.14.1
ffmpeg==1.4
filelock==3.12.0
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.2
gast==0.4.0
gdown==4.7.1
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.31
google==3.0.0
google-ai-generativelanguage==0.1.0
google-api-core==1.34.0
google-auth==1.35.0
google-auth-oauthlib==1.0.0
google-cloud-core==1.7.3
google-generativeai==0.1.0rc1
google-pasta==0.2.0
googleapis-common-protos==1.60.0
graphlib-backport==1.0.3
greenlet==2.0.2
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
h2==3.2.0
h5py==3.9.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.16.3
httplib2==0.14.0
httptools==0.6.1
httpx==0.23.2
huggingface-hub==0.14.1
humanfriendly==10.0
hyperframe==5.2.0
identify==2.5.29
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflect==7.0.0
iopath==0.1.10
ipykernel==6.22.0
ipython==8.13.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-client==8.2.0
jupyter-core==5.3.0
keras==2.13.1
keyring==18.0.1
kiwisolver==1.4.4
kubernetes==28.1.0
langchain==0.0.340
langdetect==1.0.9
langsmith==0.0.66
language-selector==0.1
launchpadlib==1.10.13
layoutparser==0.3.4
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-loader==0.3
libclang==16.0.6
librosa==0.10.1
lit==16.0.2
llvmlite==0.40.0
lockfile==0.12.2
loguru==0.7.2
louis==3.12.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.0
manifest-ml==0.0.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==3.1
nlpcloud==1.1.44
nltk==3.8.1
nodeenv==1.8.0
notion-client==2.0.0
notion2md==2.8.3
numba==0.57.0
numexpr==2.8.5
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.16.1
openai==1.2.0
opencv-python==4.8.1.78
openlm==0.0.5
openpyxl==3.1.2
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.0.1
paramiko==2.6.0
parso==0.8.3
pathtools==0.1.2
pbr==6.0.0
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
peft==0.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.4
pipreqs==0.4.13
pkgutil-resolve-name==1.3.10
platformdirs==3.5.0
pooch==1.8.0
portalocker==2.8.2
posthog==3.0.2
pre-commit==3.4.0
profanity==1.1
progressbar==2.5
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pxpx==1.0.29
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.0
pycups==1.9.73
pydantic==2.3.0
pydantic-core==2.6.3
pydeck==0.8.1b0
Pygments==2.15.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pypdf==3.17.0
pypdfium2==4.23.1
PyPika==0.48.9
pyreadline3==3.4.1
pyRFC3339==1.1
PySocks==1.7.1
pysqlite3-binary==0.5.2.post1
pytesseract==0.3.10
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-rapidjson==1.10
pytz==2023.3
pyxdg==0.26
PyYAML==6.0
pyzmq==25.0.2
rapidfuzz==2.15.1
redis==5.0.0
referencing==0.30.2
regex==2023.3.23
reportlab==3.5.34
requests==2.29.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
responses==0.18.0
rfc3986==1.5.0
rich==13.4.2
rotary-embedding-torch==0.3.6
rpds-py==0.10.3
rsa==4.9
safetensors==0.4.0
scikit-learn==1.2.2
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sentencepiece==0.1.99
sentry-sdk==1.21.1
setproctitle==1.3.2
shap==0.41.0
simplejson==3.16.0
six==1.16.0
slicer==0.0.7
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.7
SQLAlchemy==2.0.20
sqlitedict==2.1.0
sqlparse==0.4.4
ssh-import-id==5.10
stack-data==0.6.2
starlette==0.27.0
streamlit==1.27.0
sympy==1.11.1
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
testresources==2.0.1
threadpoolctl==3.1.0
tiktoken==0.5.2
timm==0.9.8
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.1.1
torchvision==0.16.0
tornado==6.3.1
TorToiSe==2.3.0
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
triton==2.0.0
tritonclient==2.34.0
typer==0.9.0
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.3.7
unstructured==0.10.28
unstructured-inference==0.7.10
unstructured.pytesseract==0.3.12
urllib3==1.26.15
usb-creator==0.3.7
uvicorn==0.23.2
uvloop==0.19.0
validators==0.22.0
virtualenv==20.24.5
wadllib==1.3.3
wandb==0.15.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.6
websocket-client==1.6.4
websockets==12.0
werkzeug==2.3.7
wrapt==1.15.0
xkit==0.0.0
xxhash==3.2.0
yarg==0.1.9
yarl==1.9.2
zipp==3.16.2
